AS.DB.API

Documentation

getCell :: Connection -> ASLocation -> IO (Maybe ASCell)

Storage Documentation

Cells key-value hashes key is produced by cellLocation (see DU.getLocationKey) and is unique fields are "cellExpression", "cellValue", "cellTags" with corresponding stringified values

DAG same as before: a set of relations access set with key DAGLocSet

Sheets stored as key (DU.getSheetKey ASSheetId) value (stringified ASSheet) additionally, the set of all locations belonging to a sheet are stored as set key (DU.getSheetSetKey ASSheetId) members (ASLocationKey) this set is updated automatically during setCells. finally, a record of all sheetKeys is stored as a set with key "sheets" and members (DU.getSheetKey sheetid)

Workbooks stored identically to Sheets

Commits stored as before, as a list of commits

Volatile locs stored as before, as a set with key volatileLocs

Cells

getCells :: Connection -> [ASLocation] -> IO [Maybe ASCell]

getCellsByKeys :: Connection -> [ByteString] -> IO [Maybe ASCell]

setCell :: Connection -> ASCell -> IO ()

setCells :: Connection -> [ASCell] -> IO ()

deleteCells :: Connection -> [ASCell] -> IO ()

deleteLocs :: Connection -> [ASLocation] -> IO ()

locationsExist :: Connection -> [ASLocation] -> IO [Bool]

getColumnCells :: Connection -> ASLocation -> IO [Maybe ASCell]

getDAG :: Connection -> IO [(ASLocation, ASLocation)]

DAG

updateDAG :: Connection -> [([ASLocation], ASLocation)] -> IO ()

updateAfterEval :: Connection -> ASUser -> ASCell -> [ASCell] -> [ASCell] -> IO ()

Commits

TODO: need to deal with large commit sizes and max number of commits

Deal with updating all DB-related things after an eval

addCommit :: Connection -> ASUser -> [ASCell] -> [ASCell] -> IO ()

Creates and pushes a commit to the DB

undo :: Connection -> IO (Maybe ASCommit)

Return a commit if possible (not possible if you undo past the beginning of time, etc) | Update the DB so that there's always a source of truth (ie we will propagate undo to all relevant users)

redo :: Connection -> IO (Maybe ASCommit)

pushCommit :: Connection -> ASCommit -> IO ()

getAllWorkbookSheets :: Connection -> IO [WorkbookSheet]

WorkbookSheets (for frontend API)

createWorkbookSheet :: Connection -> WorkbookSheet -> IO WorkbookSheet

deleteWorkbookSheet :: Connection -> WorkbookSheet -> IO ()

modifyWorkbookSheets :: Connection -> ([ASSheetId] -> [ASSheetId]) -> String -> IO ()

createWorkbook :: Connection -> [ASSheetId] -> IO ASWorkbook

Raw workbooks

getUniqueWbName :: Connection -> IO String

getWorkbook :: Connection -> String -> IO (Maybe ASWorkbook)

getAllWorkbooks :: Connection -> IO [ASWorkbook]

setWorkbook :: Connection -> ASWorkbook -> IO ()

workbookExists :: Connection -> String -> IO Bool

deleteWorkbook :: Connection -> String -> IO ()

deleteWorkbookAndSheets :: Connection -> String -> IO ()

getSheet :: Connection -> ASSheetId -> IO (Maybe ASSheet)

Raw sheets

getAllSheets :: Connection -> IO [ASSheet]

createSheet :: Connection -> ASSheet -> IO ASSheet

setSheet :: Connection -> ASSheet -> IO ()

deleteSheetUnsafe :: Connection -> ASSheetId -> IO ()

getVolatileLocs :: Connection -> IO [ASLocation]

Volatile cell methods

setChunkVolatileCells :: [ASCell] -> Redis ()

deleteChunkVolatileCells :: [ASCell] -> Redis ()

canAccessSheet :: Connection -> ASUserId -> ASSheetId -> IO Bool

Permissions

canAccess :: Connection -> ASUserId -> ASLocation -> IO Bool

canAccessAll :: Connection -> ASUserId -> [ASLocation] -> IO Bool

isPermissibleMessage :: Connection -> ASUserId -> ASMessage -> IO Bool