AS.Dispatch.Core

Documentation

runDispatchCycle :: ASUser -> MVar ServerState -> ASMessage -> IO ASMessage

Regular eval route

Go through the regular eval route

updateCell :: Connection -> ASCell -> IO (Either ASExecError ())

Eval building blocks

Takes a cell and returns an error if it tries to access a non-existant cell | Otherwise, it returns all of the immediate ancestors (used to make the lookup map)

getDescendants :: Connection -> ASCell -> IO (Either ASExecError [ASCell])

Return the descendants of a cell, which will always exist but may be locked | TODO: throw exceptions for permissions/locking

propagate :: Connection -> [ASCell] -> [ASCell] -> IO (Either ASExecError [ASCell])

Takes ancestors and descendants, create lookup map, and run eval

evalChain :: Connection -> Map ASLocation ASValue -> [ASCell] -> IO [ASCell]

Eval helpers

createListCells :: Connection -> ASLocation -> [ASValue] -> IO [ASCell]

Create a list of cells, also modify the DB for references | Not currently handling [[[]]] type things