AS.Eval.Middleware

Documentation

evalMiddleware :: ASCell -> IO ASCell

This is middleware for evaluation; we take a cell recieved with the Evaluate action tag and preprocess it

evalConnector :: ASCell -> IO ASCell

Middlewares

Change the excel expression to a python one and also possibly add a volatile tag evalInitExcel :: ASCell -> IO ASCell evalInitExcel c(Cell loc xp(Expression rawXp Excel) val ts) = do initResult <- R.evalExcel xp return $ case initResult of (Left valueE) -> Cell loc xp valueE ts (Right (newXp, isVolatile)) -> case isVolatile of False -> Cell loc newXp val ts True -> case (U.hasVolatileTag c) of True -> Cell loc newXp val ts False -> Cell loc newXp val (Volatile:ts) evalInitExcel cell = return cell