AS.Types

Documentation

type ASSheetId = Text

Sheets

data ASSheet

Constructors

Sheet

Fields

sheetId :: ASSheetId

sheetName :: String

sheetPermissions :: ASPermissions

data ASWorkbook

Constructors

Workbook

Fields

workbookName :: String

workbookSheets :: [ASSheetId]

data WorkbookSheet

Constructors

WorkbookSheet

Fields

wsName :: String

wsSheets :: [ASSheet]

data ASLocation

Core cell types

Constructors

Index

Fields

locSheetId :: ASSheetId

index :: (Int, Int)

Range

Fields

locSheetId :: ASSheetId

range :: ((Int, Int), (Int, Int))

Column

Fields

locSheetId :: ASSheetId

column :: Int

data ASValue

Constructors

NoValue

ValueNaN ()

ValueS String

ValueI Int

ValueD Double

ValueB Bool

ValueL [ASValue]

ExcelSheet

Fields

locs :: ASValue

exprs :: ASValue

vals :: ASValue

Rickshaw

Fields

rickshawData :: ASValue

ValueError

Fields

error :: String

err_type :: String

file :: String

position :: Int

ValueImage

Fields

imagePath :: String

StockChart

Fields

stockPrices :: ASValue

stockName :: String

ObjectValue

Fields

objectType :: String

jsonRepresentation :: String

StyledValue

Fields

style :: String

value :: ASValue

DisplayValue

Fields

displayValue :: String

actualValue :: ASValue

ValueE ASEvalError

data ASLangValue

Constructors

LangValue

Fields

langValue :: ASValue

lang :: ASLanguage

type ASEvalError = String

data ASLanguage

Constructors

R

Python

OCaml

CPP

Java

SQL

Excel

data ASExpression

Constructors

Expression

Fields

expression :: String

language :: ASLanguage

Reference

Fields

location :: ASLocation

referenceIndex :: (Int, Int)

data ASCellTag

Constructors

Color String

Size Int

Money

Percentage

StreamTag Stream

Tracking

Volatile

ReadOnly [ASUserId]

data ASCell

Constructors

Cell

Fields

cellLocation :: ASLocation

cellExpression :: ASExpression

cellValue :: ASValue

cellTags :: [ASCellTag]

data ExLoc

Excel

Constructors

ExSheet

Fields

name :: String

sheetLoc :: ExLoc

ExRange

Fields

first :: ExLoc

second :: ExLoc

ExIndex

Fields

d1 :: String

col :: String

d2 :: String

row :: String

data ExcelAction

Constructors

Lookup

CheckIndirectRef

LookupSheets

LookupWorkbooks

CurrentLocation

data ExcelPayload

Constructors

NoPayload

data Bloomberg

Streaming | Stream sources

Constructors

Bloomberg

Fields

url :: String

key :: String

data StreamSource

Constructors

StreamB Bloomberg

NoSource

data Stream

A stream just needs a source and a frequency

Constructors

Stream

Fields

streamSource :: StreamSource

streamFreq :: Int

data ASMessage

Message Types

Constructors

Message

Fields

messageUserId :: ASUserId

action :: ASAction

result :: ASResult

payload :: ASPayload

data ASAction

Constructors

NoAction

Acknowledge

New

Import

Open

Close

Evaluate

EvaluateRepl

Update

Get

Delete

Copy

CopyForced

Undo

Redo

Clear

UpdateWindow

AddTags

RemoveTags

data ASResult

Constructors

Success

Failure

Fields

failDesc :: String

NoResult

data QueryList

Constructors

Sheets

Workbooks

WorkbookSheets

data ASPayload

Constructors

PayloadN ()

PayloadInit ASInitConnection

PayloadDaemonInit ASInitDaemonConnection

PayloadC ASCell

PayloadCL [ASCell]

PayloadL ASLocation

PayloadLL [ASLocation]

PayloadS ASSheet

PayloadSS [ASSheet]

PayloadWB ASWorkbook

PayloadWBS [ASWorkbook]

PayloadWorkbookSheets [WorkbookSheet]

PayloadW ASWindow

PayloadU ASUserId

PayloadE ASExecError

PayloadCommit ASCommit

PayloadTags

Fields

tags :: [ASCellTag]

tagsLoc :: ASLocation

PayloadXp ASExpression

PayloadLangValue ASLangValue

PayloadList QueryList

data ASExecError

Eval Types

Constructors

Timeout

EvaluationError

Fields

evalErrorDesc :: String

DependenciesLocked

Fields

lockUserId :: ASUserId

DBNothingException

Fields

badLocs :: [ASLocation]

DBGraphUnreachable

NetworkDown

ResourceLimitReached

InsufficientPermissions

NonUniqueIdentifier

CopyNonexistentDependencies

type EitherCells = Either ASExecError [ASCell]

data ASInitConnection

Websocket types

Constructors

ASInitConnection

Fields

connUserId :: ASUserId

data ASInitDaemonConnection

Constructors

ASInitDaemonConnection

Fields

parentUserId :: ASUserId

initDaemonLoc :: ASLocation

data ASDaemon

Constructors

ASDaemon

Fields

daemonLoc :: ASLocation

daemonConn :: Connection

data ServerState

Database and state

Constructors

State

Fields

userList :: [(ASUser, [ASDaemon])]

dbConn :: Connection

data GraphQuery

Constructors

GetDescendants

GetImmediateAncestors

SetRelations

data ASWindow

Users

Constructors

Window

Fields

windowSheetId :: ASSheetId

topLeft :: (Int, Int)

bottomRight :: (Int, Int)

type ASUserId = Text

data ASUser

Constructors

User

Fields

userId :: ASUserId

userConn :: Connection

userWindows :: [ASWindow]

data ASUserGroup

Constructors

Group

Fields

groupMembers :: [ASUserId]

groupAdmins :: [ASUserId]

groupName :: Text

data ASEntity

Constructors

EntityGroup ASUserGroup

EntityUser ASUserId

data ASPermissions

Constructors

Blacklist [ASEntity]

Whitelist [ASEntity]

data ASTime

Version Control

Constructors

Time

Fields

day :: String

hour :: Int

min :: Int

sec :: Int

data ASCommit

Constructors

ASCommit

Fields

commitUserId :: ASUserId

before :: [ASCell]

after :: [ASCell]

time :: ASTime

lst :: ASValue -> [ASValue]

Convenience methods

str :: ASValue -> String

dbl :: ASValue -> Double

failureMessage :: String -> ASMessage

initialViewingWindow :: ASWindow

genericText :: Text

When sending data from server to client, the server doesn't have a userId

openPermissions :: ASPermissions